requirements.txt
telethon
rich
python-dotenv
